niedziela praca

+++ src/posts/views.py
@@ -64,32 +64,6 @@ def index(request):
 
 
 
-
-
-
-	# queryset_list = Post.objects.all() #.order_by("-timestamp")
-
-	# paginator = Paginator(queryset_list, 3)
-	# page_request_var = "page"
-	# page = request.GET.get(page_request_var)
-	# try:
-	# 	queryset = paginator.page(page)
-	# except PageNotAnInteger:
- #        # If page is not an integer, deliver first page.
-	# 	queryset = paginator.page(1)
-	# except EmptyPage:
- #        # If page is out of range (e.g. 9999), deliver last page of results.
-	# 	queryset = paginator.page(paginator.num_pages)
-
-	# context = {
-	# 	"object_list": queryset,
-	# 	"page_request_var": page_request_var
-	# }
-	# return render(request, "index.html", context)
-
-
-
-
 def posts_update(request, slug):
 	instance = get_object_or_404(Post, slug=slug)
 	form = PostForm(request.POST or None, instance=instance)
@@ -105,9 +79,24 @@ def posts_update(request, slug):
 	}
 	return render(request, "post_form.html", context)
 
+
+
 def posts_delete(request, slug):
+	instance = get_object_or_404(Post, slug=slug)
 	if not request.user.is_staff or not request.user.is_superuser:
 		return	HttpResponseRedirect("/admin/login/?next=/delete/")
-	instance = get_object_or_404(Post, slug=slug)
-	instance.delete()
-	return redirect("posts:list")
+	if request.method == "POST":
+		instance.delete()
+		return	HttpResponseRedirect("/#artykuly")
+	context = {
+		"obj": instance
+		}
+	return render(request, "delete_confirm.html", context)
+
+
+# def posts_delete(request, slug):
+# 	if not request.user.is_staff or not request.user.is_superuser:
+# 		return	HttpResponseRedirect("/admin/login/?next=/delete/")
+# 	instance = get_object_or_404(Post, slug=slug)
+# 	instance.delete()
+# 	return redirect("posts:list")
